AxiosStatus now relies on provide/inject to retrieve AxiosMonitor status info.
